uart_test: add ping test
The utility can be started in 2 flavors: client and server.
The server can accept commands from a client and send a response.
The client can send 2 commands "SEND" and "SEND_RECV". The "SEND"
command send a buffer to the server and measures the time it took to
send it. The "SEND_RECV" command sends a buffer to the server while
measuring the time and simultaneously reads a buffer having the same size.

Signed-off-by: Petre Pircalabu <petre.pircalabu@google.com>